requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
docutils==0.20.1
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mattermostautodriver==1.2.2
mmpy-bot==2.1.4
multidict==6.0.4
packaging==23.1
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
urllib3==2.0.4
yarl==1.9.2
